% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{dnorkernel}
\alias{dnorkernel}
\title{Derivative of normal distribution}
\usage{
dnorkernel(ord, X)
}
\arguments{
\item{ord}{The order of derivative.}

\item{X}{The numeric data vector.}
}
\description{
Derivative of normal distribution
}
\keyword{internal}
