% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hds}
\alias{hds}
\title{NASA Hypobaric Decompression Sickness Marker Data}
\format{
This is a data frame with 238 observations and 3 variables: L (left limit of the observed time), R (right limit of the observed time) 
        and M (marker). The marker is a score derived by combining the covariates Age, Sex, TR360, and Noadyn.
}
\usage{
data(hds)
}
\description{
This data contains the marker values with the left and right limits of the observed time for the subjects in NASA Hypobaric Decompression Sickness Data.
}
\references{
Beyene, K. M. and El Ghouch A. (2020). Time-dependent ROC curves estimator for interval-censored survival data.
}
\keyword{datasets}
