% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{integ}
\alias{integ}
\title{Numerical Integral function using Simpson's rule}
\usage{
integ(x, fx, method, n.pts = 256)
}
\arguments{
\item{x}{The numeric data vector.}

\item{fx}{The function.}

\item{method}{The character string specifying method of numerical integration. The possible options are \code{trap} for trapezoidal rule and \code{simps} for simpson'r rule.}

\item{n.pts}{Number of points.}
}
\description{
Numerical Integral function using Simpson's rule
}
\keyword{internal}
