% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{ker_dis_i}
\alias{ker_dis_i}
\title{Distribution function without the ith observation}
\usage{
ker_dis_i(X, y, wt, ktype, bw)
}
\arguments{
\item{X}{The numeric data vector.}

\item{y}{The vector where the kernel estimation is computed.}

\item{wt}{The non-negative weight vector.}

\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}".}

\item{bw}{A numeric bandwidth value.}
}
\value{
Returns the estimated value for the bandwith parameter.
}
\description{
Distribution function without the ith observation
}
\author{
Kassu Mehari Beyene  and Anouar El Ghouch
}
\keyword{internal}
