% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{kfunc}
\alias{kfunc}
\title{Function to evaluate the matrix of data vector minus the grid points divided by the bandwidth value.}
\usage{
kfunc(ktype = "normal", difmat)
}
\arguments{
\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}". By default, the "\code{normal}" kernel is used.}

\item{difmat}{A numeric matrix of sample data (X) minus evaluation points (x0) divided by bandwidth value (bw).}
}
\value{
Returns the matrix resulting from evaluating \code{difmat}.
}
\description{
Function to evaluate the matrix of data vector minus the grid points divided by the bandwidth value.
}
\keyword{internal}
