% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{kfunction}
\alias{kfunction}
\title{Kernel distribution function}
\usage{
kfunction(ktype, X)
}
\arguments{
\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}".}

\item{X}{A numeric vector of sample data.}
}
\value{
Returns a vector resulting from evaluating X.
}
\description{
Kernel distribution function
}
\keyword{internal}
