% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mayo}
\alias{mayo}
\title{Mayo Marker Data}
\format{
A data frame with 312 observations and 4 variables: time (event time/censoring time), censor (censoring
       indicator), mayoscore4, mayoscore5. The two scores are derived from 4 and 5 covariates
       respectively.
}
\usage{
data(mayo)
}
\description{
Two marker values with event time and censoring status for the subjects in Mayo PBC data.
}
\references{
Heagerty, P. J., and Zheng, Y. (2005). Survival model predictive accuracy and ROC curves. \emph{Biometrics}, 61(1), 92-105.
}
\keyword{datasets}
