% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{muro}
\alias{muro}
\title{The value of squared integral x^2 k(x) dx and integral x k(x) K(x) dx}
\usage{
muro(ktype)
}
\arguments{
\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}".}
}
\description{
The value of squared integral x^2 k(x) dx and integral x k(x) K(x) dx
}
\keyword{internal}
