% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{NR}
\alias{NR}
\title{The normal reference bandwidth selection for weighted data}
\usage{
NR(X, wt, ktype = "normal")
}
\arguments{
\item{X}{The numeric data vector.}

\item{wt}{The non-negative weight vector.}

\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}". By default, the "\code{normal}" kernel is used.}
}
\value{
Returns the computed value for the bandwith parameter.
}
\description{
This function computes the data-driven bandwidth for smoothing the ROC (or distribution) function. It is an extension of the classical (unweighted) normal reference bandwith selection method for the case of weighted data.
 If all the weights are equal, then the value is the same as the one obtained using the classical normal reference method.
}
\details{
This mehod is derived from the asymptotic mean square error of the time-dependent smoothed ROC function.
It require estimating the unknown quantity \eqn{\kappa (ROC)=\int_{-\infty}^{\infty} \left\{ ROC^{''}(u) \right\} ^2 du}, where \eqn{ROC^{''}(.)} is the second derivative of the ROC function.
To estimate this it is asumed that the random variable follows a normal distribution and the weighted nature is accounted in the estimation of the unknown
parameters of the assumed normal distribution.
For smoothing the ROC function the following kernel can be considered: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}".
}
\examples{
library(cenROC)

X <- rnorm(100) # random data vector
wt <- runif(100) # weight vector

## Normal reference bandwidth selection
bandNR <- NR(X = X, wt = wt)

bandNR$bw
}
\references{
Beyene, K. M. and El Ghouch A. (2020). Smoothed time-dependent ROC curves for right-censored survival data.
}
\author{
Kassu Mehari Beyene, Catholic University of Louvain. \code{<kasu.beyene@uclouvain.be>}

Anouar El Ghouch, Catholic University of Louvain. \code{<anouar.elghouch@uclouvain.be>}
}
