% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{PI}
\alias{PI}
\title{The plug-in bandwidth selection for weighted data}
\usage{
PI(X, wt, ktype = "normal")
}
\arguments{
\item{X}{The numeric vector of random variable.}

\item{wt}{The non-negative weight vector.}

\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}". By default, the "\code{normal}" kernel is used.}
}
\value{
Returns the computed value for the bandwith parameter.
}
\description{
This function computes the data-driven bandwidth for smoothing the ROC (or distribution) function. It is an extension of the classical (unweighted) direct plug-in bandwith selection method for the case of weighted data.
 If all the weights are equal, then the value is the same as the one obtained using the classical plug-in method.
}
\details{
This mehod is derived from the asymptotic mean square error of the time-dependent smoothed ROC function.
It is the modified version of the plug-in bandwidth selection method for the kernel smoothed distribution function estimation.
The quantity \eqn{\kappa (ROC)=\int_{-\infty}^{\infty} \left\{ ROC^{''}(u) \right\} ^2 du}, where \eqn{ROC^{''}(.)} is the second derivative of the ROC function,
is estimated nonparametriclly using the relation \eqn{\kappa(ROC)= - E\left\{ \mathcal{W} ROC^{(3)}(Z)  \right\}}, where \eqn{ROC^{(3)}(.)} is the third derivative of the ROC function .
For the pilot bandwidth needed for estimating the quantity \eqn{E\left\{ \mathcal{W} ROC^{(3)}(Z)  \right\}}, the normal reference method is used.
}
\examples{
library(cenROC)

X <- rnorm(100) # random data vector
wt <- runif(100) # weight vector

## Plug-in bandwidth selection
bandPI <- PI(X = X, wt = wt)

bandPI$bw
}
\references{
Beyene, K. M. and El Ghouch A. (2020). Smoothed time-dependent ROC curves for right-censored survival data.
}
\author{
Kassu Mehari Beyene, Catholic University of Louvain. \code{<kasu.beyene@uclouvain.be>}

Anouar El Ghouch, Catholic University of Louvain. \code{<anouar.elghouch@uclouvain.be>}
}
