% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{RocFun}
\alias{RocFun}
\title{ROC estimation function}
\usage{
RocFun(U, D, M, bw = "NR", method, ktype)
}
\arguments{
\item{U}{The vector of grid points where the ROC curve is estimated.}

\item{D}{The event indicator.}

\item{M}{The numeric vector of marker values for which the time-dependent ROC curves is computed.}

\item{bw}{The bandwidth parameter for smoothing the ROC function. The possible options are \code{NR} normal reference method; \code{PI} plug-in method and \code{CV} cross-validation method. The default is the \code{NR} normal reference method.}

\item{method}{is the method of ROC curve estimation. The possible options are \code{emp} emperical metod; \code{untra} smooth without boundary correction and \code{tra} is smooth ROC curve estimation with boundary correction.}

\item{ktype}{A character string giving the type kernel to be used: "\code{normal}", "\code{epanechnikov}", "\code{biweight}", or "\code{triweight}".}
}
\description{
ROC estimation function
}
\references{
Beyene, K. M. and El Ghouch A. (2019). Smoothed time-dependent ROC curves for right-censored survival data. <\url{https://dial.uclouvain.be/pr/boreal/object/boreal:219643}>.
}
\author{
Beyene K. Mehari and El Ghouch Anouar
}
\keyword{internal}
