% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{wIQR}
\alias{wIQR}
\title{Weighted inter-quartile range estimation}
\usage{
wIQR(X, wt)
}
\arguments{
\item{X}{The numeric data vector.}

\item{wt}{The non-negative weight vector.}
}
\description{
Weighted inter-quartile range estimation
}
\keyword{internal}
