% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{wquantile}
\alias{wquantile}
\title{Weighted quartile estimation}
\usage{
wquantile(X, wt, p = 0.5)
}
\arguments{
\item{X}{The numeric data vector.}

\item{wt}{The non-negative weight vector.}

\item{p}{The percentile value. The defult is 0.5.}
}
\description{
Weighted quartile estimation
}
\keyword{internal}
