% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{wvar}
\alias{wvar}
\title{Weighted variance estimation}
\usage{
wvar(X, wt, na.rm = FALSE)
}
\arguments{
\item{X}{The numeric data vector.}

\item{wt}{The non-negative weight vector.}

\item{na.rm}{The character indicator wether to consider missing value(s) or not. The defult is FALSE.}
}
\description{
Weighted variance estimation
}
\keyword{internal}
